Removes base role class, makes role class more efficient
